+++ demo/demo/urls.py
@@ -1,7 +1,6 @@
-from distutils.version import StrictVersion as version
-import django
 from django.conf.urls import include, url
 from django.contrib import admin
+from django.urls import path
 
 from demo.views import FormPreview, DemoValidateViewFromSchema
 
@@ -14,14 +13,5 @@ urlpatterns = [
     url(r'^preview/(?P<pk>\d+)/', FormPreview.as_view()),
     url(r'^forms/',
         include(('demo.builder.urls', 'builder'), namespace='builder')),
+    path(r'admin/', admin.site.urls, 'admin'),
 ]
-
-if version(django.get_version()) < version("2.0"):
-    urlpatterns += [
-        url(r'^admin/', include(admin.site.urls)),
-    ]
-else:
-    from django.urls import path  # noqa
-    urlpatterns += [
-        path(r'admin/', admin.site.urls, 'admin'),
-    ]