+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ ChangeLog
 master (unreleased)
 ===================
 
+- Drop support for Django 1.11 (#398, #395).
 - Drop support for Django REST Framework 3.8 (#382).
-- Ensure ``jsonfield`` compatibility check. Documentation is amended, and an ``ImportWarning`` is thrown as soon as you're using Django 1.11 (#395).
 - Fix Postgresql configuration in CircleCI regarding the authentication (#395).
 - Small cleanups of Python2-related code.